+++ src/components/[guild]/CreatePoap/index.ts
@@ -1,0 +1,3 @@
+import CreatePoap from "./CreatePoap"
+
+export default CreatePoap
