--- src/components/[guild]/CreatePoapButton/index.ts
@@ -1,3 +1,0 @@
-import CreatePoapButton from "./CreatePoapButton"
-
-export default CreatePoapButton
