+++ src/components/[guild]/EditGuild/components/DeleteGuildButton/index.ts
