+++ src/components/[guild]/EditGuild/index.ts
@@ -1,0 +1,3 @@
+import EditGuild from "./EditGuild"
+
+export default EditGuild
