--- src/components/[guild]/EditGuildButton/index.ts
@@ -1,3 +1,0 @@
-import EditGuildButton from "./EditGuildButton"
-
-export default EditGuildButton
