+++ src/components/[guild]/GuildMenu/index.ts
@@ -1,0 +1,3 @@
+import GuildMenu from "./GuildMenu"
+
+export default GuildMenu
